% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamicTable.R
\name{addColumn_Server}
\alias{addColumn_Server}
\title{Server logic of module for adding a column}
\usage{
addColumn_Server(input, output, session, data)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}

\item{data}{source data as data.frame}
}
\value{
reactive data.frame of modified source data
}
\description{
Server logic of module for adding a column
}
