% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamicTable.R
\name{addColumn_UI}
\alias{addColumn_UI}
\title{UI elements of module for adding a column}
\usage{
addColumn_UI(id, label = "Add column")
}
\arguments{
\item{id}{module id}

\item{label}{label of the button}
}
\value{
UI module
}
\description{
UI elements of module for adding a column
}
