% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamicTable.R
\name{addRow_UI}
\alias{addRow_UI}
\title{UI elements of module for adding a row}
\usage{
addRow_UI(id, label = "Add")
}
\arguments{
\item{id}{module id}

\item{label}{label of the button}
}
\value{
UI module
}
\description{
UI elements of module for adding a row
}
