% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulesResourceButtons.R
\name{addRowRc_UI}
\alias{addRowRc_UI}
\title{UI elements of Resource tab module for adding a row}
\usage{
addRowRc_UI(id, label = "Add")
}
\arguments{
\item{id}{module id}

\item{label}{label of the button}
}
\value{
UI module
}
\description{
UI elements of Resource tab module for adding a row
}
