% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbpManager-pkg.R
\docType{package}
\name{cbpManager-pkg}
\alias{cbpManager-pkg}
\title{cbpManager}
\description{
`cbpManager` is an R package that provides an interactive Shiny-based
 graphical user interface for...
}
\author{
Arsenij Ustjanzew \email{arsenij.ustjanzew@gmail.com}
}
