% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbpManager.R
\name{cbpManager}
\alias{cbpManager}
\title{Launch cbpManager}
\usage{
cbpManager(studyDir = NULL, logDir = NULL, returnAppObj = FALSE, ...)
}
\arguments{
\item{studyDir}{Path to study folder containing studies of cBioPortal.}

\item{logDir}{Path where a logfile should be saved. If NULL, logs are
not stored}

\item{returnAppObj}{Logical value, whether to return the app object if set to
TRUE. Default behavior: directly runs the app (FALSE)}

\item{...}{Further parameters that are used by \code{shiny::runApp}, e.g.
\code{host} or \code{port}.}
}
\value{
shiny application object
}
\description{
Launch the cbpManager Shiny application.
}
\examples{
if (interactive()) {
  cbpManager()
}
}
