% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{check_input_dates}
\alias{check_input_dates}
\title{Check the input of dates}
\usage{
check_input_dates(diagnosisDate, startDate = NULL, endDate = NULL)
}
\arguments{
\item{diagnosisDate}{date of first diagnosis}

\item{startDate}{start date of timeline event}

\item{endDate}{end date of timeline event}
}
\value{
Returns a number indicating the warning
}
\description{
Check the input of dates
}
