% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{config_colUI}
\alias{config_colUI}
\title{Function to generate an HTML Shiny input widget string based on the definition of the config file.}
\usage{
config_colUI(
  colname,
  id_prefix,
  data = NULL,
  mode = c("add", "edit"),
  selected_row = NULL,
  textvalue = NULL
)
}
\arguments{
\item{colname}{The column name.}

\item{id_prefix}{ID for the input widget generated in function 'generateUIwidgets'.}

\item{data}{Data.frame}

\item{mode}{"add" or "edit" - whether to use existing values or not.}

\item{selected_row}{The selected row of 'data'.}
}
\value{
A column specific shiny UI-widget (HTML string)
}
\description{
Function to generate an HTML Shiny input widget string based on the definition of the config file.
}
\examples{
cbpManager:::writeLogfile(tempdir(), "data_clinical_patient.txt")

}
