% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{convertDataFrame}
\alias{convertDataFrame}
\title{Convert the data.frame to the appropriate file format for cBioPortal}
\usage{
convertDataFrame(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
Data.frame formated for the cBioPortal file format
}
\description{
Convert the data.frame to the appropriate file format for cBioPortal
}
