% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_name}
\alias{create_name}
\title{Sanitize names}
\usage{
create_name(x, toupper = TRUE)
}
\arguments{
\item{x}{A character string.}

\item{toupper}{If TRUE, the name wil be upper-case; if FALSE, the name will be lower-case.}
}
\value{
A sanitized string.
}
\description{
This function takes a character string, replaces spaces by underscores and runs make.names.
}
\examples{
cbpManager:::create_name("Study name 1") #returns "STUDY_NAME_1"
cbpManager:::create_name("FANCY;name", toupper = FALSE) #returns "fancy.name"

}
