% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamicTable.R
\name{deleteColumn_Server}
\alias{deleteColumn_Server}
\title{Server logic of module for deleting a column}
\usage{
deleteColumn_Server(input, output, session, data, exclude)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}

\item{data}{source data as data.frame}

\item{exclude}{column names that shoud be excluded from deletion}
}
\value{
reactive data.frame of modified source data
}
\description{
Server logic of module for deleting a column
}
