% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamicTable.R
\name{deleteColumn_UI}
\alias{deleteColumn_UI}
\title{UI elements of module for deleting a column}
\usage{
deleteColumn_UI(id, label = "Delete column(s)")
}
\arguments{
\item{id}{module id}

\item{label}{label of the button}
}
\value{
UI module
}
\description{
UI elements of module for deleting a column
}
