% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamicTable.R
\name{deleteRow_Server}
\alias{deleteRow_Server}
\title{Server logic of module for removing a row}
\usage{
deleteRow_Server(input, output, session, data, selected_row)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}

\item{data}{source data as data.frame}

\item{selected_row}{Index of the selected row from the table}
}
\value{
reactive data.frame of modified source data
}
\description{
Server logic of module for removing a row
}
