% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamicTable.R
\name{deleteRow_UI}
\alias{deleteRow_UI}
\title{UI elements of module for removing a row}
\usage{
deleteRow_UI(id, label = "Delete")
}
\arguments{
\item{id}{module id}

\item{label}{label of the button}
}
\value{
UI module
}
\description{
UI elements of module for removing a row
}
