% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulesResourceButtons.R
\name{deleteRowRc_Server}
\alias{deleteRowRc_Server}
\title{Server logic of Resource tab module for deleting a row}
\usage{
deleteRowRc_Server(
  input,
  output,
  session,
  data,
  selected_row,
  mode = "default",
  sample_data = NULL,
  patient_data = NULL,
  study_data = NULL
)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}

\item{data}{source data as data.frame}

\item{selected_row}{Index of the selected row from the table}

\item{mode}{TODOTODO}

\item{sample_data}{TODOTODO}

\item{patient_data}{TODOTODO}

\item{study_data}{TODOTODO}
}
\value{
reactive data.frame of modified source data
}
\description{
Server logic of Resource tab module for deleting a row
}
