% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulesResourceButtons.R
\name{deleteRowRc_UI}
\alias{deleteRowRc_UI}
\title{UI elements of module for removing a row}
\usage{
deleteRowRc_UI(id, label = "Delete")
}
\arguments{
\item{id}{Module id}

\item{label}{Label of the button}
}
\value{
UI module
}
\description{
UI elements of module for removing a row
}
