% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.create_name}
\alias{.create_name}
\title{Sanitize component names}
\usage{
.create_name(x)
}
\arguments{
\item{x}{A character string.}
}
\value{
A sanitized string.
}
\description{
This function takes a character string, replaces spaces by underscores and runs make.names.
}
