% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamicTable.R
\name{editRow_Server}
\alias{editRow_Server}
\title{Server logic of module for editing a row}
\usage{
editRow_Server(
  input,
  output,
  session,
  data,
  patient_ids = NULL,
  dates_first_diagnosis = NULL,
  selected_row = NULL,
  mode = c("timeline", "timepoint")
)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}

\item{data}{source data as data.frame}

\item{patient_ids}{reactive vector of existing patient IDs}

\item{dates_first_diagnosis}{data.frame with dates of the first diagnosis and patient IDs}

\item{selected_row}{the index of the selected row}

\item{mode}{Mode of the timeline data. Controls which columns are displayed.}
}
\value{
reactive data.frame of modified source data
}
\description{
Server logic of module for editing a row
}
