% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamicTable.R
\name{editRow_UI}
\alias{editRow_UI}
\title{UI elements of module for editing a row}
\usage{
editRow_UI(id, label = "Edit")
}
\arguments{
\item{id}{module id}

\item{label}{label of the button}
}
\value{
UI module
}
\description{
UI elements of module for editing a row
}
