% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulesResourceButtons.R
\name{editRowRc_UI}
\alias{editRowRc_UI}
\title{UI elements of module for editing a row}
\usage{
editRowRc_UI(id, label = "Edit")
}
\arguments{
\item{id}{Module id}

\item{label}{Label of the button}
}
\value{
UI module
}
\description{
UI elements of module for editing a row
}
