% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{findDatatype}
\alias{findDatatype}
\title{Define the datatype required in cBioPortal file format}
\usage{
findDatatype(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
String according to the data type of the column
}
\description{
Define the datatype required in cBioPortal file format
}
