% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fncols}
\alias{fncols}
\title{Add empty column to a data.frame, if column does not exist in the data.frame}
\usage{
fncols(data, cname)
}
\arguments{
\item{data}{data.frame}

\item{cname}{column name}
}
\value{
data.frame
}
\description{
Add empty column to a data.frame, if column does not exist in the data.frame
}
