% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generateOncotreeUIwidgets}
\alias{generateOncotreeUIwidgets}
\title{tbd}
\usage{
generateOncotreeUIwidgets(colname, mode = c("add", "edit"))
}
\arguments{
\item{colname}{tbd}

\item{mode}{tbd}
}
\value{
tbd
}
\description{
tbd
}
