% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generateTimelineUI}
\alias{generateTimelineUI}
\title{tbd}
\usage{
generateTimelineUI(
  colname,
  mode = c("add", "edit"),
  data = NULL,
  selected_row = NULL,
  patientIDs = NULL
)
}
\arguments{
\item{colname}{tbd}

\item{mode}{tbd}

\item{data}{tbd}

\item{selected_row}{tbd}

\item{patientIDs}{tbd}
}
\value{
tbd
}
\description{
tbd
}
