% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getSampleIDs}
\alias{getSampleIDs}
\title{Get Sample IDs associated with Patient IDs from the data_clinical_sample.txt file}
\usage{
getSampleIDs(file_path, patIDs)
}
\arguments{
\item{file_path}{A character string.}

\item{patIDs}{A character string.}
}
\value{
vector with Sample IDs
}
\description{
Get Sample IDs associated with Patient IDs from the data_clinical_sample.txt file
}
\examples{
cbpManager:::getSampleIDs(
    system.file("study/testpatient/data_clinical_sample.txt", package = "cbpManager"), 
    patIDs = "Testpatient")

}
