% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{importPatientData}
\alias{importPatientData}
\title{Import patient data into current study data.frames}
\usage{
importPatientData(
  mode = c("patient", "sample", "mutations", "timelines"),
  file_name,
  file_path,
  patIDs,
  data,
  associatedSampleIDs = NULL
)
}
\arguments{
\item{mode}{tbd}

\item{file_name}{tbd}

\item{file_path}{tbd}

\item{patIDs}{tbd}

\item{data}{tbd}

\item{associatedSampleIDs}{tbd}
}
\value{
a data.frame
}
\description{
Import patient data into current study data.frames
}
