% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{IsDate}
\alias{IsDate}
\title{Check if input is in the appropriate date format}
\usage{
IsDate(mydate, date.format = "\%Y-\%m-\%d")
}
\arguments{
\item{date.format}{string describig the date format}

\item{df}{date}
}
\description{
Check if input is in the appropriate date format
}
