% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{launchApp}
\alias{launchApp}
\title{Launch cbpManager}
\usage{
launchApp(studyDir = NULL, ...)
}
\arguments{
\item{studyDir}{Path to study folder containing studies of cBioPortal.}

\item{...}{Further parameters that are used by \code{shiny::runApp}, e.g.
\code{host} or \code{port}.}
}
\value{
shiny application object
}
\description{
Launch the cbpManager Shiny application.
}
\examples{
if ( interactive() ) {
  launchApp()
}

}
