% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_meta}
\alias{read_meta}
\title{Read a metafile line by line into a data frame.}
\usage{
read_meta(filepath)
}
\arguments{
\item{filepath}{Filepath with filename of metafile.}
}
\value{
data.frame
}
\description{
Read a metafile line by line into a data frame.
}
\examples{
read_meta(system.file("study", "testpatient", "meta_CNA.txt", package = "cbpManager"))

}
