% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulesResourceButtons.R
\name{saveResource_Server}
\alias{saveResource_Server}
\title{Server logic of module for saving the resource data}
\usage{
saveResource_Server(
  input,
  output,
  session,
  data,
  study_id,
  data_filename,
  meta_filename,
  resource_type = c("SAMPLE", "DEFINITION", "PATIENT", "STUDY")
)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}

\item{data}{source data as data.frame}

\item{study_id}{the current study ID}

\item{data_filename}{TODOTODO}

\item{meta_filename}{TODOTODO}

\item{resource_type}{TODOTODO}
}
\value{
nothing to return
}
\description{
Server logic of module for saving the resource data
}
