% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamicTable.R
\name{saveTimeline_Server}
\alias{saveTimeline_Server}
\title{Server logic of module for saving the source data}
\usage{
saveTimeline_Server(input, output, session, data, study_id)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}

\item{data}{source data as data.frame}

\item{study_id}{the current study ID}
}
\value{
boolean value; TRUE if function was used.
}
\description{
Server logic of module for saving the source data
}
