% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamicTable.R
\name{saveTimeline_UI}
\alias{saveTimeline_UI}
\title{UI elements of module for saving the data}
\usage{
saveTimeline_UI(id, label = "Save")
}
\arguments{
\item{id}{module id}

\item{label}{label of the button}
}
\value{
UI module
}
\description{
UI elements of module for saving the data
}
