% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{setupConda_cbpManager}
\alias{setupConda_cbpManager}
\title{Install conda environment with basilisk before launching the app}
\usage{
setupConda_cbpManager()
}
\value{
Nothing to return
}
\description{
Install conda environment with basilisk before launching the app
}
\examples{
\dontrun{
setupConda_cbpManager()
}

}
