% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{updateOncotreeUIwidgets}
\alias{updateOncotreeUIwidgets}
\title{Generate UI input widget}
\usage{
updateOncotreeUIwidgets(session, row_last_clicked, mode = c("add", "edit"))
}
\arguments{
\item{mode}{"add" or "edit" - wether to use existing values or not}

\item{colname}{A character string - the name of the column, that will be the label of the input}

\item{tab}{"Patient", "Sample" - The used tab; sets the html id prefix of the input}

\item{data}{A data.frame.}

\item{selected_row}{A number indicating the row number of the selected row in the data.frame.}
}
\value{
A sanitized string.
}
\description{
Generate UI input widget
}
