% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulesResourceButtons.R
\name{validateResourcePatient}
\alias{validateResourcePatient}
\title{Validate resource_patient input}
\usage{
validateResourcePatient(values)
}
\arguments{
\item{values}{List of input values}
}
\value{
boolean
}
\description{
Validate resource_patient input
}
