% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulesResourceButtons.R
\name{validateResourceSample}
\alias{validateResourceSample}
\title{Validate resource_sample input}
\usage{
validateResourceSample(values)
}
\arguments{
\item{values}{List of input values}
}
\value{
boolean
}
\description{
Validate resource_sample input
}
