% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulesResourceButtons.R
\name{validateResourceStudy}
\alias{validateResourceStudy}
\title{Validate resource_study input}
\usage{
validateResourceStudy(values)
}
\arguments{
\item{values}{List of input values}
}
\value{
boolean
}
\description{
Validate resource_study input
}
