% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basilisk.R
\docType{data}
\name{ValidationDependencies}
\alias{ValidationDependencies}
\alias{.validation_dependencies}
\title{Validation Dependencies}
\format{
A character vector containing the pinned versions of all Python packages on
which the import validation depends.
}
\usage{
.validation_dependencies
}
\description{
Vector defining a set of Python dependencies and versions required to operate
with the validation scripts for cBioPortal
}
\keyword{datasets}
