% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{writeLogfile}
\alias{writeLogfile}
\title{Write a line to a logfile containing the date, time, username (from Shinyproxy), and the name of the modified file.}
\usage{
writeLogfile(outdir, modified_file, log_filename = "cbpManager_logfile.txt")
}
\arguments{
\item{outdir}{directory, where the logfile should be saved}

\item{modified_file}{Name of the modified file}

\item{log_filename}{Name of the logfile}
}
\value{
Nothing to return
}
\description{
Write a line to a logfile containing the date, time, username (from Shinyproxy), and the name of the modified file.
}
\examples{
cbpManager:::writeLogfile(tempdir(), "data_clinical_patient.txt")

}
