% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulesResourceButtons.R
\name{addRowRc_Server}
\alias{addRowRc_Server}
\title{Server logic of Resource tab module for adding a row}
\usage{
addRowRc_Server(
  input,
  output,
  session,
  data,
  patient_ids = NULL,
  sample_ids = NULL,
  resource_ids = NULL,
  resource_type = c("definition", "sample", "patient", "study")
)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}

\item{data}{Source data as data.frame}

\item{patient_ids}{Reactive vector of existing patient IDs}

\item{sample_ids}{Reactive data.frame of existing patient IDs and sample IDs}

\item{resource_ids}{Reactive data.frame of data_resource_definition}

\item{resource_type}{The type of the resource.
Can be "definition", "sample", "patient", "study"}
}
\value{
reactive data.frame of modified source data
}
\description{
Server logic of Resource tab module for adding a row
}
