% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generateOncotreeUIwidgets}
\alias{generateOncotreeUIwidgets}
\title{Create shiny UI-widget for specific columns of oncotree entries}
\usage{
generateOncotreeUIwidgets(colname, mode = c("add", "edit"))
}
\arguments{
\item{colname}{column name}

\item{mode}{determines the inputId prefix of the UI-widget}
}
\value{
A oncotree specific shiny UI-widget
}
\description{
Create shiny UI-widget for specific columns of oncotree entries
}
