% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generateTimelineUI}
\alias{generateTimelineUI}
\title{generate UI elements for the
generate UI elements for timeline modal dialog}
\usage{
generateTimelineUI(
  colname,
  mode = c("add", "edit"),
  data = NULL,
  selected_row = NULL,
  patientIDs = NULL
)
}
\arguments{
\item{data}{data.frame}

\item{cname}{column name}
}
\value{
data.frame
}
\description{
generate UI elements for the
generate UI elements for timeline modal dialog
}
