% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{importPatientData}
\alias{importPatientData}
\title{Import patient data into current study data.frames}
\usage{
importPatientData(
  mode = c("patient", "sample", "mutations", "timelines"),
  file_name,
  file_path,
  patIDs,
  data,
  associatedSampleIDs = NULL
)
}
\arguments{
\item{mode}{Defines the type of imported data}

\item{file_name}{Filename of source data}

\item{file_path}{Filepath with filename of source data}

\item{patIDs}{PATIENT_IDs of patients that should be imported}

\item{data}{Source data, to be subsetted according to patIDs}

\item{associatedSampleIDs}{The sample IDs associated to the patIDs}
}
\value{
data.frame
}
\description{
Import patient data into current study data.frames
}
