% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{IsDate}
\alias{IsDate}
\title{Check if input is in the appropriate date format}
\usage{
IsDate(mydate, date.format = "\%Y-\%m-\%d")
}
\arguments{
\item{mydate}{date}

\item{date.format}{string describig the date format}
}
\value{
boolean
}
\description{
Check if input is in the appropriate date format
}
\examples{
cbpManager:::IsDate("2020-02-20")
cbpManager:::IsDate("20.01.2020", date.format = "\%d.\%m.\%Y")

}
