% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulesResourceButtons.R
\name{saveResource_Server}
\alias{saveResource_Server}
\title{Server logic of module for saving the resource data}
\usage{
saveResource_Server(
  input,
  output,
  session,
  data,
  study_id,
  data_filename,
  meta_filename,
  resource_type = c("SAMPLE", "DEFINITION", "PATIENT", "STUDY")
)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}

\item{data}{Source data as data.frame}

\item{study_id}{The current study ID}

\item{data_filename}{File name of the data file}

\item{meta_filename}{file name of the meta file}

\item{resource_type}{The type of the resource.
Can be "definition", "sample", "patient", "study"}
}
\value{
boolean value; TRUE if function was used.
}
\description{
Server logic of module for saving the resource data
}
