% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{updateOncotreeUIwidgets}
\alias{updateOncotreeUIwidgets}
\title{tbd}
\usage{
updateOncotreeUIwidgets(session, row_last_clicked, mode = c("add", "edit"))
}
\arguments{
\item{session}{tbd}

\item{row_last_clicked}{tbd}

\item{mode}{tbd}
}
\value{
tbd
}
\description{
tbd
}
