% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{updateOncotreeUIwidgets}
\alias{updateOncotreeUIwidgets}
\title{Updates UI-widgets for specific columns of oncotree entries}
\usage{
updateOncotreeUIwidgets(
  session,
  row_last_clicked,
  mode = c("add", "edit"),
  tab = c("Patient", "Sample")
)
}
\arguments{
\item{session}{Shiny session}

\item{row_last_clicked}{the index of the row last clicked in the oncotree_table}

\item{mode}{determines the inputId prefix of the UI-widget}

\item{tab}{"Patient", "Sample" - The used tab; sets the html id pr}
}
\value{
nothing to return
}
\description{
Updates UI-widgets for specific columns of oncotree entries
}
