% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulesResourceButtons.R
\name{validateResourceDefinition}
\alias{validateResourceDefinition}
\title{Validate resource_definition input}
\usage{
validateResourceDefinition(values, resourceDf, mode = "add")
}
\arguments{
\item{values}{List of input values}

\item{resourceDf}{data.frame of data_resource_definition}

\item{mode}{The mode of the function ('add' or 'edit')}
}
\value{
boolean
}
\description{
Validate resource_definition input
}
